projects/ActionsUptime/web/requirements.txt
requests
python-whois
sentry-sdk